requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.2
dj-database-url==2.1.0
Django==4.2.9
django-browser-reload==1.12.1
django-on-heroku==1.1.2
django-rest-framework==0.1.0
django-tailwind==3.6.0
djangorestframework==3.14.0
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.2.0
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
zipp==3.17.0
